+++ xarray/core/groupby.py
@@ -362,6 +362,22 @@ class GroupBy(object):
                     del obj.coords[dim]
         return obj
 
+    def _maybe_stack(self, applied):
+        """
+        This constructs MultiIndex if 'applied' does not have self._group_dim.
+        It may happen if a single item is selected from MultiIndex-ed array.
+        """
+        if not hasattr(self._group, 'to_index'):
+            return applied
+        index = self._group.to_index()
+        if not isinstance(index, pd.MultiIndex):
+            return applied
+        else:
+            return [ds if self._group_dim in ds.coords
+                    else ds.expand_dims(index.names).stack(
+                        **{self._group.name: index.names})
+                    for ds in applied]
+
     def fillna(self, value):
         """Fill missing values in this object by group.
 
@@ -528,6 +544,7 @@ class DataArrayGroupBy(GroupBy, ImplementsArrayReduce):
         """Recombine the applied objects like the original."""
         applied_example, applied = peek_at(applied)
         coord, dim, positions = self._infer_concat_args(applied_example)
+        applied = self._maybe_stack(applied)
         if shortcut:
             combined = self._concat_shortcut(applied, dim, positions)
         else:
@@ -619,6 +636,7 @@ class DatasetGroupBy(GroupBy, ImplementsDatasetReduce):
     def _combine(self, applied):
         """Recombine the applied objects like the original."""
         applied_example, applied = peek_at(applied)
+        applied = self._maybe_stack(applied)
         coord, dim, positions = self._infer_concat_args(applied_example)
         combined = concat(applied, dim)
         combined = _maybe_reorder(combined, dim, positions)