+++ xarray/core/indexing.py
@@ -265,14 +265,18 @@ def get_dim_indexers(data_obj, indexers):
 
 def remap_label_indexers(data_obj, indexers, method=None, tolerance=None):
     """Given an xarray data object and label based indexers, return a mapping
-    of equivalent location based indexers. Also return a mapping of updated
-    pandas index objects (in case of multi-index level drop).
+    of equivalent location based indexers.
+    In case of multi-index level drop, it also returns
+    (new_indexes) a mapping of updated pandas index objects and
+    (selected_dims) a mapping from the original dims to selected (dropped)
+    dims.
     """
     if method is not None and not isinstance(method, str):
         raise TypeError('``method`` must be a string')
 
     pos_indexers = {}
     new_indexes = {}
+    selected_dims = {}
 
     dim_indexers = get_dim_indexers(data_obj, indexers)
     for dim, label in iteritems(dim_indexers):
@@ -291,8 +295,15 @@ def remap_label_indexers(data_obj, indexers, method=None, tolerance=None):
             pos_indexers[dim] = idxr
             if new_idx is not None:
                 new_indexes[dim] = new_idx
-
-    return pos_indexers, new_indexes
+                if isinstance(new_idx, pd.MultiIndex):
+                    selected_dims[dim] = [name for name in index.names
+                                          if name not in new_idx.names]
+                else:
+                    selected_dims[dim] = [name for name in index.names
+                                          if name != new_idx.name]
+            if isinstance(idxr, int) and idxr in (0, 1):
+                selected_dims[dim] = index.names
+    return pos_indexers, new_indexes, selected_dims
 
 
 def slice_slice(old_slice, applied_slice, size):