+++ xarray/tests/test_groupby.py
@@ -23,6 +23,18 @@ def test_consolidate_slices():
         _consolidate_slices([slice(3), 4])
 
 
+def test_multi_index_groupby_apply_dataarray():
+    # regression test for GH873
+    ds = xr.DataArray(np.random.randn(3, 4), dims=['x', 'y'],
+                      coords={'x': ['a', 'b', 'c'], 'y': [1, 2, 3, 4]})
+    doubled = 2 * ds
+    group_doubled = (ds.stack(space=['x', 'y'])
+                     .groupby('space')
+                     .apply(lambda x: 2 * x)
+                     .unstack('space'))
+    assert doubled.equals(group_doubled)
+
+
 def test_multi_index_groupby_apply():
     # regression test for GH873
     ds = xr.Dataset({'foo': (('x', 'y'), np.random.randn(3, 4))},
@@ -70,5 +82,5 @@ def test_groupby_duplicate_coordinate_labels():
     actual = array.groupby('x').sum()
     assert expected.equals(actual)
 
-    
+
 # TODO: move other groupby tests from test_dataset and test_dataarray over here